updates to exif package
